% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{create_pop_growth_params}
\alias{create_pop_growth_params}
\title{Create population growth parameters.}
\usage{
create_pop_growth_params(t_crit = 300, t1 = 100, N_0 = 1e+06,
  initial_reproduction_number = 0.8, gamma = 24, mu = 1/70, beta1 = 0,
  period = 1, rho = 0.1, S_0 = 0.99, I_0 = 1e-06, R_0 = 0,
  eta = 1/N_0)
}
\arguments{
\item{t_crit}{the time at which the the expected value of the
basic reproduction number should equal 1}

\item{t1}{the time at which the the population size starts to grow.}

\item{N_0}{the initial population size}

\item{initial_reproduction_number}{the initial value of the basic
reproduction number}

\item{gamma}{the recovery rate}

\item{mu}{the death rate}

\item{beta1}{the amplitude of seasonal variation in the transmission rate}

\item{period}{the period of seasonal variation in the transmission rate}

\item{rho}{the binomial probability of a case being reported}

\item{S_0}{the initial weight of susceptibles}

\item{I_0}{the initial weight of infectives}

\item{R_0}{the initial weight of recovereds}

\item{eta}{the rate of infection from outside of the population}
}
\value{
a named vector suitable for the "params" argument to the
\code{create_sir_simulator}
}
\description{
This function calculates parameters consistent with the population
starting to grow at a given time point at a constant rate such
that the expected value of the basic reproduction number crosses
one from below at a later given time point. It does this by
setting the transmission rate, beta, and the rate of change of the
birth rate, alpha, accordingly.
}
\examples{

params <- create_pop_growth_params()
foo <- create_sir_simulator()
out <- pomp::simulate(foo, times=seq(0, 400), params=params)
out <- as(out, "data.frame")

opar <- par(mfrow=c(3, 1))
plot(R~time, data=out, type='l')
plot(N~time, data=out, type='l')
repn <- out$N * params["beta0"] / (params["gamma"] + params["mu"])
plot(out$time, repn, type='l')
par(opar)

}
\seealso{
\code{\link{create_sir_simulator}}
}

