% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{create_sir_simulator}
\alias{create_sir_simulator}
\title{Create SIR gillespie simulator}
\usage{
create_sir_simulator()
}
\description{
Create SIR gillespie simulator
}
\examples{

# See strong seasonal forcing
foo <- create_sir_simulator()
params <- structure(c(gamma=24, mu=0.01, eta=10e-06,
                      beta0=11e-06, beta1=.99,
                      period=1, t1=10, alpha=0,
                      rho=0.1, S_0=0.9974, I_0=0,
                      R_0=1 - .9974, N_0=1e+06))
out <- pomp::simulate(foo, times=seq(0, 20, by=1/26), params=params)
out <- as(out, "data.frame")

opar <- par(mfrow=c(4, 1))
plot((S/N)~time, data=out, type="l")
plot(cases~time, data=out, type="l")
repn <- params["beta0"] * params["N_0"] / (params["mu"] + params["gamma"])
repn <- repn * (1 + params["beta1"] * sinpi (2 * out$time / params["period"]))
plot(out$time, repn, type='l')
plot(out$cases, repn)
par(opar)

}

